Implement answer checking and use of tempDB
